pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "resume-builder-agent"
version = "0.1.0"
description = "AI Resume Builder with ATS Optimization"
requires-python = ">=3.11"
dependencies = [
    # LangGraph
    "langgraph>=0.2.28",
    "langgraph-checkpoint-sqlite>=2.0.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    # Database
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.0",
    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0",
    "email-validator>=2.0.0",
    # FastAPI
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "firebase-admin>=6.2.0",
    "pydantic>=2.0.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "pytesseract>=0.3.10",
    "pillow>=10.0.0",
    "pdf2image>=1.16.3",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]